requirements.txt
zyndai-agent[heartbeat,ngrok]
langchain
langchain-openai
langchain-community
langchain-classic
flask
flask-cors
requests
python-dotenv
